refactor and organize lib/ folder
