+++ lib/view/drafts/Users.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:hive_app/utils/ColorPalette.dart';
 
-import 'package:hive_app/view/pages/UserCard.dart'; 
+import 'package:hive_app/view/drafts/UserCard.dart';
 import 'package:hive_app/view_model/user.vm.dart';
 import 'package:hive_app/models/user.model.dart';
 import 'package:provider/provider.dart';
 import 'package:hive_app/data/remote/response/Status.dart';
-import 'package:hive_app/utils/ColorPalette.dart';
 
 class Users extends StatefulWidget {
   static const String id = "home_screen";
@@ -52,7 +51,7 @@ class _UsersScreenState extends State<Users> {
             case Status.COMPLETED:
               print("Log :: COMPLETED");
               return _getUsersListView(viewModel.userModel.data?.users);
-              // return _getUsersListView(viewModel.userModel.data?.users);
+            // return _getUsersListView(viewModel.userModel.data?.users);
             default:
               return Container();
           }
@@ -70,5 +69,4 @@ class _UsersScreenState extends State<Users> {
           );
         });
   }
-
 }
