+++ lib/view/pages/Calendar.dart
@@ -4,7 +4,7 @@ import 'package:hive_app/utils/ColorPalette.dart';
 import 'package:hive_app/view/widgets/EventList.dart';
 import 'package:hive_app/view_model/event.vm.dart';
 import 'package:hive_app/data/remote/response/Status.dart';
-import 'package:hive_app/view/pages/ViewsHeader.dart';
+import 'package:hive_app/view/widgets/ViewsHeader.dart';
 import 'package:hive_app/services/notification_services.dart';
 import 'package:provider/provider.dart';
